+++ CHANGELOG.md
@@ -1,8 +1,28 @@
 # Release Notes
 
-## [Unreleased](https://github.com/inertiajs/inertia-laravel/compare/v2.0.12...2.x)
+## [Unreleased](https://github.com/inertiajs/inertia-laravel/compare/v2.0.14...2.x)
 
 - Nothing!
+
+## [v2.0.14](https://github.com/inertiajs/inertia-laravel/compare/v2.0.13...v2.0.14) - 2025-12-10
+
+### What's Changed
+
+* Exclude loaded once props from deferred metadata by [@pascalbaljet](https://github.com/pascalbaljet) in https://github.com/inertiajs/inertia-laravel/pull/796
+
+**Full Changelog**: https://github.com/inertiajs/inertia-laravel/compare/v2.0.13...v2.0.14
+
+## [v2.0.13](https://github.com/inertiajs/inertia-laravel/compare/v2.0.12...v2.0.13) - 2025-12-10
+
+### What's Changed
+
+* Include `onceProps` metadata for `fresh` props by [@skryukov](https://github.com/skryukov) in https://github.com/inertiajs/inertia-laravel/pull/794
+
+### New Contributors
+
+* [@skryukov](https://github.com/skryukov) made their first contribution in https://github.com/inertiajs/inertia-laravel/pull/794
+
+**Full Changelog**: https://github.com/inertiajs/inertia-laravel/compare/v2.0.12...v2.0.13
 
 ## [v2.0.12](https://github.com/inertiajs/inertia-laravel/compare/v2.0.11...v2.0.12) - 2025-12-09